Move other logic to ViewModel.
