+++ LaTeXSierotkaRemover/MainWindow.xaml
@@ -8,12 +8,15 @@
         xmlns:conv="clr-namespace:LaTeXSierotkaRemover.Converters"
         mc:Ignorable="d"
         Title="MainWindow" Height="450" Width="800">
+    
     <Window.DataContext>
         <vm:MainViewModel x:Name="ViewModel" />
     </Window.DataContext>
+    
     <Window.Resources>
         <conv:BooleanToVisibilityConverter x:Key="BooleanToVisibilityConverter" />
     </Window.Resources>
+    
     <Grid>
         <StackPanel>
             <Grid>
@@ -32,9 +35,9 @@
                     <ColumnDefinition />
                 </Grid.ColumnDefinitions>
                 <Label Grid.Column="0" Content="'Sierotki' base: " />
-                <TextBox x:Uid="UrlBaseTextBox" x:Name="UrlLatexTextBox" Grid.Column="1" Text="./index.tex" Margin="2" VerticalAlignment="Stretch" HorizontalAlignment="Stretch" />
+                <TextBox x:Uid="UrlBaseTextBox" x:Name="UrlLatexTextBox" Grid.Column="1" Text="{Binding DocumentAddress, Mode=TwoWay}" Margin="2" VerticalAlignment="Stretch" HorizontalAlignment="Stretch" />
             </Grid>
-            <Button x:Uid="DoThis" x:Name="DoThis" Content="Remove 'Sierotki'" Click="DoThis_Click" />
+            <Button x:Uid="DoThis" x:Name="DoThis" Content="Remove 'Sierotki'" Command="{Binding RemoveSierotkiCommand}" />
         </StackPanel>
         <Grid x:Uid="LoadingLockGrid" Visibility="{Binding IsVisibleLoadingLockGrid, Mode=TwoWay, Converter={StaticResource BooleanToVisibilityConverter}}" x:Name="LoadingLockGrid" Background="#99555555">
             <TextBlock Text="PLEASE WAIT..." HorizontalAlignment="Center" VerticalAlignment="Center"/>
